refactor(windows): preserve lifecycle module boundaries
