+++ crates/astrid-cli/src/commands/daemon.rs
@@ -12,6 +12,13 @@ use crate::bootstrap::find_companion_binary;
 use crate::commands::{daemon_control, daemon_process};
 use crate::{socket_client, theme};
 
+#[path = "daemon_runtime.rs"]
+mod runtime;
+#[cfg(test)]
+use runtime::{ShutdownRequestOutcome, shutdown_request_outcome, stop_confirmation};
+pub(crate) use runtime::{StopConfirmation, handle_stop};
+use runtime::{is_handshake_rejection, remove_runtime_files_if_unowned};
+
 const DAEMON_READY_TIMEOUT_SECS: u64 = 60;
 const DAEMON_READY_POLL_MILLIS: u64 = 50;
 const DAEMON_READY_POLL: Duration = Duration::from_millis(DAEMON_READY_POLL_MILLIS);
@@ -27,19 +34,6 @@ pub(crate) struct StartArgs {
     /// Keep the daemon attached to this command and propagate its exit.
     #[arg(long)]
     pub(crate) foreground: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum StopConfirmation {
-    ConfirmedGone,
-    Unconfirmed,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-enum ShutdownRequestOutcome {
-    Acknowledged,
-    Escalate(String),
-    Rejected(String),
 }
 
 const fn readiness_attempts(timeout_secs: u64, poll_millis: u64) -> u64 {
@@ -706,386 +700,6 @@ fn status_response(response: KernelResponse) -> Result<DaemonStatus> {
             anyhow::bail!("daemon rejected status request: {message}")
         },
         other => anyhow::bail!("daemon returned an unexpected status response: {other:?}"),
-    }
-}
-
-/// Handle `astrid stop`.
-///
-/// A shutdown request over the socket only earns an ACK ("shutting down"), not a
-/// guarantee the process exited and released the singleton/state-db lock. So we
-/// capture the recorded PID BEFORE asking, then confirm the process actually
-/// exits — escalating with platform-native forced termination if it wedges
-/// mid-shutdown — before reporting success. Runtime files (transport,
-/// readiness, PID) are removed only once the
-/// daemon is confirmed gone; if a kill can't confirm exit, they are LEFT so
-/// `astrid start`/`restart` still see the recorded PID and give an actionable
-/// message instead of failing on the held lock with a raw DB error.
-pub(crate) async fn handle_stop() -> Result<StopConfirmation> {
-    let socket_path = socket_client::try_proxy_socket_path()?;
-    let pid_path = socket_client::try_pid_path()?;
-
-    // Capture the daemon's identity up front: it deletes its own PID file only
-    // on a CLEAN exit, so reading it before shutdown is the only reliable way to
-    // keep a handle for confirming exit / terminating a wedged shutdown.
-    let recorded = daemon_control::read_pid_file(&pid_path)?;
-    let endpoint_reachable = daemon_endpoint_reachable(&socket_path).await;
-
-    // Genuinely nothing running: no socket AND no live recorded process.
-    let recorded_alive = recorded
-        .as_ref()
-        .is_some_and(|identity| daemon_control::is_process_alive(identity.pid));
-    if !endpoint_reachable && !recorded_alive {
-        return Ok(report_confirmed_cleanup(
-            "No Astrid daemon is running.",
-            &pid_path,
-            &socket_path,
-        ));
-    }
-
-    // Graceful path: the socket is present and serviceable.
-    // Deliberately bypass the selected-workspace check: stopping a daemon is
-    // the recovery path when that daemon belongs to another project/layout.
-    let recovery_client = if endpoint_reachable {
-        match tokio::time::timeout(
-            Duration::from_secs(5),
-            socket_client::connect_kernel_for_recovery(),
-        )
-        .await
-        {
-            Ok(Ok(client)) => Some(client),
-            Ok(Err(error)) if is_handshake_rejection(&error) => {
-                return Err(error)
-                    .context("daemon rejected the authenticated lifecycle-recovery connection");
-            },
-            Ok(Err(error)) => {
-                eprintln!(
-                    "{}",
-                    theme::Theme::warning(&format!(
-                        "Daemon recovery connection failed ({error:#}); escalating through the \
-                         recorded process identity."
-                    ))
-                );
-                None
-            },
-            Err(_) => {
-                eprintln!(
-                    "{}",
-                    theme::Theme::warning(
-                        "Daemon recovery connection timed out; escalating through the recorded \
-                         process identity."
-                    )
-                );
-                None
-            },
-        }
-    } else {
-        None
-    };
-
-    if let Some(client) = recovery_client {
-        let mut client = client.with_timeout(Duration::from_secs(10));
-        match shutdown_request_outcome(
-            client
-                .request(KernelRequest::Shutdown {
-                    reason: Some("astrid stop".to_string()),
-                })
-                .await,
-        ) {
-            ShutdownRequestOutcome::Acknowledged => {
-                // ACK only — confirm the process actually exits before
-                // declaring success, and escalate if it wedged.
-                return Ok(confirm_graceful_stop(recorded, &pid_path, &socket_path).await);
-            },
-            ShutdownRequestOutcome::Escalate(reason) => {
-                eprintln!(
-                    "{}",
-                    theme::Theme::warning(&format!(
-                        "Authenticated daemon shutdown failed ({reason}); escalating through the \
-                        recorded process identity."
-                    ))
-                );
-            },
-            ShutdownRequestOutcome::Rejected(reason) => anyhow::bail!("{reason}"),
-        }
-    }
-
-    // Orphan path: the socket is present but unreachable (hung/half-dead
-    // daemon), OR the socket is already gone but a live recorded daemon is still
-    // holding the lock. A clean shutdown request is impossible either way, so
-    // terminate the recorded PID (identity-gated) and clean up. Using the PID we
-    // captured up front — not a re-read — closes the window where the daemon
-    // deletes its own PID file mid-wedge.
-    let outcome = match &recorded {
-        Some(identity) => daemon_control::terminate_known(identity).await,
-        None => daemon_control::KillOutcome::NotRunning,
-    };
-    // A shutdown may have reached the daemon even when its response was lost.
-    // If no process remains to terminate, refresh the transport observation:
-    // an absent endpoint confirms shutdown, while a connected/indeterminate
-    // endpoint may belong to another live daemon and must remain fail-closed.
-    let endpoint_reachable = if matches!(outcome, daemon_control::KillOutcome::NotRunning) {
-        daemon_endpoint_reachable(&socket_path).await
-    } else {
-        endpoint_reachable
-    };
-    Ok(report_orphan_stop(
-        outcome,
-        endpoint_reachable,
-        &pid_path,
-        &socket_path,
-    ))
-}
-
-fn is_handshake_rejection(error: &anyhow::Error) -> bool {
-    error
-        .downcast_ref::<astrid_uplink::socket_client::HandshakeRejected>()
-        .is_some()
-}
-
-async fn daemon_endpoint_reachable(socket_path: &Path) -> bool {
-    match astrid_core::local_transport::connect_outcome(socket_path).await {
-        Ok(astrid_core::local_transport::ConnectOutcome::Connected(stream)) => {
-            drop(stream);
-            true
-        },
-        Ok(
-            astrid_core::local_transport::ConnectOutcome::Absent
-            | astrid_core::local_transport::ConnectOutcome::Stale,
-        ) => false,
-        Err(_) => true,
-    }
-}
-
-fn shutdown_request_outcome<E>(
-    response: std::result::Result<KernelResponse, E>,
-) -> ShutdownRequestOutcome
-where
-    E: std::fmt::Display,
-{
-    match response {
-        Ok(KernelResponse::Success(_)) => ShutdownRequestOutcome::Acknowledged,
-        Ok(KernelResponse::Error(reason)) => {
-            ShutdownRequestOutcome::Rejected(format!("daemon rejected shutdown: {reason}"))
-        },
-        Ok(other) => ShutdownRequestOutcome::Rejected(format!(
-            "daemon returned an unexpected shutdown response: {other:?}"
-        )),
-        Err(error) => ShutdownRequestOutcome::Escalate(format!("shutdown request failed: {error}")),
-    }
-}
-
-/// After a graceful shutdown ACK, confirm the daemon process actually exited —
-/// an ACK is "shutting down", not "exited and released the lock". Wait for the
-/// recorded PID to die; if it wedges past the grace window it is still holding
-/// the lock, so escalate through the same identity-gated termination path as
-/// an unreachable orphan. Runtime files are cleaned only once the process is
-/// gone.
-async fn confirm_graceful_stop(
-    recorded: Option<daemon_control::DaemonIdentity>,
-    pid_path: &Path,
-    socket_path: &Path,
-) -> StopConfirmation {
-    let Some(identity) = recorded else {
-        // Legacy pidless daemon (or an unresolved PID): an ACK does not prove
-        // exit. The daemon owns cleanup; without a recorded generation we must
-        // not unlink a newer daemon's identity or report confirmed success.
-        eprintln!(
-            "{}",
-            theme::Theme::warning(
-                "The daemon acknowledged shutdown, but its process identity was unavailable; \
-                 exit could not be confirmed and runtime state was left intact."
-            )
-        );
-        return StopConfirmation::Unconfirmed;
-    };
-    let pid = identity.pid;
-
-    if daemon_control::wait_for_exit(pid, daemon_control::GRACE).await {
-        return report_confirmed_cleanup("Astrid daemon stopped.", pid_path, socket_path);
-    }
-
-    // Acknowledged but still alive past the grace window → wedged mid-shutdown,
-    // still holding the lock. Escalate with forced termination
-    // (identity-gated).
-    eprintln!(
-        "{}",
-        theme::Theme::warning(
-            "Daemon acknowledged shutdown but is still running; escalating with forced \
-             termination so the state-db lock is released."
-        )
-    );
-    let outcome = daemon_control::terminate_known(&identity).await;
-    report_orphan_stop(outcome, true, pid_path, socket_path)
-}
-
-/// Report a forced-stop outcome and clean up runtime files ONLY when the
-/// daemon is confirmed gone. When a kill can't confirm exit (`StillAlive` /
-/// `Unverified`), the socket/PID files are LEFT in place so `astrid start` /
-/// `astrid restart` still see the recorded PID and can print an actionable
-/// message rather than failing on the held lock with a raw DB error.
-fn report_orphan_stop(
-    outcome: daemon_control::KillOutcome,
-    endpoint_reachable: bool,
-    pid_path: &Path,
-    socket_path: &Path,
-) -> StopConfirmation {
-    match &outcome {
-        daemon_control::KillOutcome::NotRunning => {
-            if endpoint_reachable {
-                eprintln!(
-                    "{}",
-                    theme::Theme::warning(
-                        "The daemon transport is reachable, but authenticated shutdown failed and \
-                         no daemon process could be verified. Leaving runtime state intact."
-                    )
-                );
-            }
-        },
-        daemon_control::KillOutcome::TermExited | daemon_control::KillOutcome::KilledExited => {},
-        daemon_control::KillOutcome::StillAlive => {
-            eprintln!(
-                "{}",
-                theme::Theme::error(
-                    "An unresponsive Astrid daemon did not exit even after forced termination; the \
-                     state-db lock may still be held. Inspect the process before retrying."
-                )
-            );
-        },
-        daemon_control::KillOutcome::Unverified(pid) => {
-            eprintln!(
-                "{}",
-                theme::Theme::warning(&format!(
-                    "A process may hold the recorded daemon PID {pid}, but I can't confirm either \
-                     its absence or that it's the Astrid daemon — not killing it. If the daemon is \
-                     genuinely stuck, inspect PID {pid} and stop it manually."
-                ))
-            );
-        },
-    }
-    let confirmation = stop_confirmation(outcome, endpoint_reachable);
-    if confirmation == StopConfirmation::ConfirmedGone {
-        let message = if matches!(outcome, daemon_control::KillOutcome::NotRunning) {
-            "No Astrid daemon is running."
-        } else {
-            "Stopped an unresponsive Astrid daemon."
-        };
-        return report_confirmed_cleanup(message, pid_path, socket_path);
-    }
-    confirmation
-}
-
-fn report_confirmed_cleanup(
-    success_message: &str,
-    pid_path: &Path,
-    socket_path: &Path,
-) -> StopConfirmation {
-    match remove_runtime_files_if_unowned(pid_path, socket_path) {
-        Ok(true) => {
-            println!("{}", theme::Theme::success(success_message));
-            StopConfirmation::ConfirmedGone
-        },
-        Ok(false) => {
-            eprintln!(
-                "{}",
-                theme::Theme::warning(
-                    "The stopped daemon exited, but another daemon generation now owns the \
-                     runtime namespace. Its files were left intact."
-                )
-            );
-            StopConfirmation::Unconfirmed
-        },
-        Err(error) => {
-            eprintln!(
-                "{}",
-                theme::Theme::warning(&format!(
-                    "The stopped daemon exited, but runtime cleanup could not be verified: {error}. \
-                     Runtime files were left for a later recovery."
-                ))
-            );
-            StopConfirmation::Unconfirmed
-        },
-    }
-}
-
-/// Remove runtime artifacts only while holding the daemon namespace lock.
-///
-/// A newly booting generation acquires this same lock before publishing any
-/// shared runtime state. `Ok(false)` therefore means another generation owns
-/// the namespace and no path may be removed.
-fn remove_runtime_files_if_unowned(pid_path: &Path, socket_path: &Path) -> std::io::Result<bool> {
-    let run_dir = pid_path.parent().ok_or_else(|| {
-        std::io::Error::new(
-            std::io::ErrorKind::InvalidInput,
-            "daemon PID path has no run directory",
-        )
-    })?;
-    let root = run_dir.parent().ok_or_else(|| {
-        std::io::Error::new(
-            std::io::ErrorKind::InvalidInput,
-            "daemon run directory has no Astrid home",
-        )
-    })?;
-    let home = astrid_core::dirs::AstridHome::from_path(root);
-    if home.pid_path() != pid_path || home.socket_path() != socket_path {
-        return Err(std::io::Error::new(
-            std::io::ErrorKind::InvalidInput,
-            "daemon runtime paths do not belong to one Astrid home",
-        ));
-    }
-    let Some(_singleton) = astrid_core::platform_fs::try_acquire_daemon_singleton(&home)? else {
-        return Ok(false);
-    };
-
-    astrid_core::local_transport::remove_endpoint(socket_path)?;
-    remove_file_if_present(&home.ready_path())?;
-    remove_file_if_present(&home.token_path())?;
-    remove_file_if_present(pid_path)?;
-    Ok(true)
-}
-
-fn remove_file_if_present(path: &Path) -> std::io::Result<()> {
-    match std::fs::remove_file(path) {
-        Ok(()) => Ok(()),
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => Ok(()),
-        Err(error) => Err(error),
-    }
-}
-
-/// Whether a stop outcome CONFIRMS the daemon is gone — the only condition under
-/// which runtime files (transport, readiness, PID, token) may be removed.
-///
-/// This is the crux of the wedge fix (#1120): `StillAlive`/`Unverified` mean a
-/// process may still hold the state-db lock, so the files are kept, leaving the
-/// recorded PID for `astrid start`/`restart` to find and act on rather than
-/// racing a fresh daemon onto a held lock (which surfaces as a raw "Database …
-/// is already locked" error). Pure over its input so the invariant is testable
-/// without a live daemon. Takes the `Copy` outcome by value (trivially small).
-const fn stop_confirmed_gone(outcome: daemon_control::KillOutcome) -> bool {
-    matches!(
-        outcome,
-        daemon_control::KillOutcome::NotRunning
-            | daemon_control::KillOutcome::TermExited
-            | daemon_control::KillOutcome::KilledExited
-    )
-}
-
-/// Combine process evidence with transport evidence.
-///
-/// `NotRunning` proves absence only when the transport is also unreachable. A
-/// reachable endpoint whose authenticated recovery handshake failed could
-/// still belong to the daemon, and without a verified PID there is no safe
-/// destructive fallback. Keep that state unconfirmed so restart and update
-/// cannot replace a potentially live owner.
-const fn stop_confirmation(
-    outcome: daemon_control::KillOutcome,
-    endpoint_reachable: bool,
-) -> StopConfirmation {
-    if endpoint_reachable && matches!(outcome, daemon_control::KillOutcome::NotRunning) {
-        StopConfirmation::Unconfirmed
-    } else if stop_confirmed_gone(outcome) {
-        StopConfirmation::ConfirmedGone
-    } else {
-        StopConfirmation::Unconfirmed
     }
 }
